% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkcso.R
\name{cso_message}
\alias{cso_message}
\title{Display a message in the console}
\usage{
cso_message(msg)
}
\arguments{
\item{msg}{character. The message to display.}
}
\value{
No return value. Prints a message to the console if allowed.
}
\description{
This function prints a message to the console if the package option
\code{cso_silent} is set to 0. If \code{cso_silent} is 1 (the default),
the message will be suppressed.
}
\details{
This helper function is part of the internal messaging system of the CSO
hydrological model package. It is mainly used to provide user feedback
during model execution or data processing steps.
}
\examples{
# Display a message (if silent mode is disabled)
cso_defaults(silent = 0)
cso_message("Model run started")

# No message is displayed if silent mode is enabled
cso_defaults(silent = 1)
cso_message("This will not be shown")

}
\seealso{
\code{\link[=cso_defaults]{cso_defaults()}}
}
